concept of virtual DOM and how state of the object changes and that triggers react to update real DOM
Virtual DOM: light weight in-memory representation of the UI,
Automatic restarting of browser when we save is called hot module reloading.
the object (<h1>hello world</h1> we see is part of the virtual DOM)
when the state of the object changes, react will get a new react element and compare this with the previous one, reach out to the real dom and then it will update it.